moved contents of keypair module to model module.
